requirements.txt
Flask==3.0.2
qrcode==7.4.2
Pillow==10.2.0
removebg==0.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
